Move some files into common util folder
