+++ src/common/common/util/string_data_mapper.cpp
@@ -1,0 +1,11 @@
+#include "string_data_mapper.h"
+
+void StringDataMapper::parse(const std::string& data)
+{
+
+}
+
+std::string StringDataMapper::findValue(const std::string& key)
+{
+    return std::string();
+}
