+++ src/server/network/server.cpp
@@ -6,7 +6,7 @@
 #include <SFML/System/Clock.hpp>
 #include <algorithm>
 #include <common/debug.h>
-#include <common/obd_parser.h>
+#include <common/util/obd_parser.h>
 #include <iostream>
 #include <thread>
